Revert "refactor(cloudflare): improve McpAgent constraint abstraction (#517)"
This reverts commit 84d4fe8bc90eb28f1b0b2507442fb4be45df2597.
